refactor: fold speculative output stats into speculative_worker_impl.
Move SpeculativeOutputStats and its per-position accounting from a
dedicated common header into speculative_worker_impl, next to the other
speculative helpers. This drops the standalone speculative_metrics
namespace, the extra header target, and the :common dependency added
for distributed_runtime/worker_service. Template instantiations now
live in a single TU instead of every includer.
